app/requirements.txt
Flask==3.0.3
Flask_Cors==4.0.0
google_api_python_client==2.126.0
keras==2.10.0
Keras_Preprocessing==1.1.2
numpy==1.24.4
python-dotenv==1.0.1
tensorflow==2.10.1